Update README.md
This paper helps to solve one problem in phytopathology and two problems in medical research. The first problem is retinal vascular segmentation for diagnosing fundus diseases. Unfortunately,
the existing methods of retinal vascular segmentation have some problems including wrong segmentation of pathological information and low microvascular segmentation. The second problem is to diagnose pigmented skin lesions including melanoma in complex images. The issue with lesions image classification is that the segmentation does not identify correctly the lesions region and the features extracted are not very pertinent. The third problem is that, it is difficult to segment and diagnose Alternaria disease and Leaf‑minear pest on tomato leaves due to the similarity of their colored areas in the early and middle stages of their outbreak. Therefore, to solve the first problem, this research has used DRIVE database (digital retinal images for vessel extraction). To unravel the second problem, which is the diagnosis of skin lesions including melanoma, image dataset from the Atlas Dermoscopy has been used. Finally, this research has used a unique database consisting of 172 images of tomato leaves infected with both types of Alternaria disease and Leaf‑minear pest and 100 healthy tomato leaves to untie the third hurdle. For the purpose of targeted segmentation, the images in this database have been collected in the early and middle stages of outbreak of the Alternaria disease and Leaf‑minear pest on tomato leaves. This paper has developed and utilized an Efficient semantic segmentation neural network
(E‑SCNN) to solve three foregoing problems. The proposed method has used Softmax classifier for classifing these three problems. The test of proposed method in RGB, YCBCR and HSV color spaces showed that this method has good accuracy and speed in all three color spaces for Retinal vascular segmentation of diagnosing fundus diseases, and for skin lesions segmentation, as
well as for segmentation of Alternaria disease and Leaf‑minear pest. RGB color space had better performance than YCBCR and HSV color spaces for solving three foregoing problems. Results
showed that the proposed method achieved an accuracy of 96.91Results also demonstrated that the proposed method achieved an accuracy of 99.25Additionally, the results indicated that the proposed method achieved an accuracy of 96.13Proposed method compared to the best previous
method on DRIVE database proved that accuracy has increased by In addition, proposed method compared to the best previous method on the Atlas Dermoscopy database, showed that
accuracy has increased by Equally, in comparison, proposed method has Proposed E‑SCNN has submitted a type of deep learning architecture that has high accuracy and speed even in datasets with a small number of images. Proposed E‑SCNN has low sensitivity to the light intensity of images, and has practical application in solving various medical and phytopathology problems and many other problems as well.